requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
alembic
aiofiles
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.1
httpx
celery[redis]
redis
